# public/assets/api.yaml
openapi: 3.0.0
paths:
...
  /sysOauthClients:
    get:
      tags:
        - SysOauthClient
      summary: 'Find list SysOauthClient'
      description: 'Returns list of SysOauthClient'
      operationId: getSysOauthClient
      parameters:
        -
          name: search
          in: query
          description: 'search by column defined'
          schema:
            type: object
        -
          name: order
          in: query
          description: 'order by column defined'
          schema:
            type: object
        -
          name: page
          in: query
          description: 'page to show'
          schema:
            type: int32
        -
          name: limit
          in: query
          description: 'count data display per page'
          schema:
            type: int32
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/SysOauthClient' } }
                  pagination: { properties: { currentPage: { type: integer }, totalPage: { type: integer } }, type: object }
                type: object
            application/xml:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/SysOauthClient' } }
                  pagination: { type: array, items: { $ref: '#/components/schemas/SysOauthClient' } }
                type: object
        '404':
          description: 'SysOauthClient not found'
      security:
        -
          bearer_auth: []
    post:
      tags:
        - SysOauthClient
      summary: 'Add a new SysOauthClient to the store'
      operationId: addSysOauthClient
      requestBody:
        $ref: '#/components/requestBodies/SysOauthClient'
      responses:
        '201':
          description: 'Created SysOauthClient'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SysOauthClient'
            application/xml:
              schema:
                $ref: '#/components/schemas/SysOauthClient'
        '405':
          description: 'Invalid input'
      security:
        -
          bearer_auth: []
  '/sysOauthClients/{id}':
    get:
      tags:
        - SysOauthClient
      summary: 'Find SysOauthClient by ID'
      description: 'Returns a single SysOauthClient'
      operationId: getSysOauthClientById
      parameters:
        -
          name: id
          in: path
          description: 'ID of SysOauthClient to return'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SysOauthClient'
            application/xml:
              schema:
                $ref: '#/components/schemas/SysOauthClient'
        '400':
          description: 'Invalid ID supplier'
        '404':
          description: 'SysOauthClient not found'
      security:
        -
          bearer_auth: []
    put:
      tags:
        - SysOauthClient
      summary: 'Update an existing SysOauthClient'
      operationId: updateSysOauthClient
      parameters:
        -
          name: id
          in: path
          description: 'SysOauthClient id to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: '#/components/requestBodies/SysOauthClient'
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'SysOauthClient not found'
        '405':
          description: 'Validation exception'
      security:
        -
          bearer_auth: []
    delete:
      tags:
        - SysOauthClient
      summary: 'Deletes a SysOauthClient'
      operationId: deleteSysOauthClient
      parameters:
        -
          name: id
          in: path
          description: 'SysOauthClient id to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Pet not found'
      security:
        -
          bearer_auth: []
  /sysPermissions:
    get:
      tags:
        - SysPermission
      summary: 'Find list SysPermission'
      description: 'Returns list of SysPermission'
      operationId: getSysPermission
      parameters:
        -
          name: search
          in: query
          description: 'search by column defined'
          schema:
            type: object
        -
          name: order
          in: query
          description: 'order by column defined'
          schema:
            type: object
        -
          name: page
          in: query
          description: 'page to show'
          schema:
            type: int32
        -
          name: limit
          in: query
          description: 'count data display per page'
          schema:
            type: int32
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/SysPermission' } }
                  pagination: { properties: { currentPage: { type: integer }, totalPage: { type: integer } }, type: object }
                type: object
            application/xml:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/SysPermission' } }
                  pagination: { type: array, items: { $ref: '#/components/schemas/SysPermission' } }
                type: object
        '404':
          description: 'SysPermission not found'
      security:
        -
          bearer_auth: []
    post:
      tags:
        - SysPermission
      summary: 'Add a new SysPermission to the store'
      operationId: addSysPermission
      requestBody:
        $ref: '#/components/requestBodies/SysPermission'
      responses:
        '201':
          description: 'Created SysPermission'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SysPermission'
            application/xml:
              schema:
                $ref: '#/components/schemas/SysPermission'
        '405':
          description: 'Invalid input'
      security:
        -
          bearer_auth: []
  '/sysPermissions/{id}':
    get:
      tags:
        - SysPermission
      summary: 'Find SysPermission by ID'
      description: 'Returns a single SysPermission'
      operationId: getSysPermissionById
      parameters:
        -
          name: id
          in: path
          description: 'ID of SysPermission to return'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SysPermission'
            application/xml:
              schema:
                $ref: '#/components/schemas/SysPermission'
        '400':
          description: 'Invalid ID supplier'
        '404':
          description: 'SysPermission not found'
      security:
        -
          bearer_auth: []
    put:
      tags:
        - SysPermission
      summary: 'Update an existing SysPermission'
      operationId: updateSysPermission
      parameters:
        -
          name: id
          in: path
          description: 'SysPermission id to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: '#/components/requestBodies/SysPermission'
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'SysPermission not found'
        '405':
          description: 'Validation exception'
      security:
        -
          bearer_auth: []
    delete:
      tags:
        - SysPermission
      summary: 'Deletes a SysPermission'
      operationId: deleteSysPermission
      parameters:
        -
          name: id
          in: path
          description: 'SysPermission id to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Pet not found'
      security:
        -
          bearer_auth: []
...
components:
  schemas:
...
    SysOauthClient:
      title: SysOauthClient
      description: SysOauthClient
      properties:
        client_id:
          nullable: false
          title: client_id
          description: client_id
          type: string
          format: '-'
          maxLength: 256
        resource_ids:
          nullable: true
          title: resource_ids
          description: resource_ids
          type: string
          format: '-'
          maxLength: 256
        client_secret:
          nullable: true
          title: client_secret
          description: client_secret
          type: string
          format: '-'
          maxLength: 256
        scope:
          nullable: true
          title: scope
          description: scope
          type: string
          format: '-'
          maxLength: 256
        authorized_grant_types:
          nullable: true
          title: authorized_grant_types
          description: authorized_grant_types
          type: string
          format: '-'
          maxLength: 256
        web_server_redirect_uri:
          nullable: true
          title: web_server_redirect_uri
          description: web_server_redirect_uri
          type: string
          format: '-'
          maxLength: 256
        authorities:
          nullable: true
          title: authorities
          description: authorities
          type: string
          format: '-'
          maxLength: 256
        access_token_validity:
          nullable: true
          title: access_token_validity
          description: access_token_validity
          type: integer
          format: '-'
          maxLength: 11
        refresh_token_validity:
          nullable: true
          title: refresh_token_validity
          description: refresh_token_validity
          type: integer
          format: '-'
          maxLength: 11
        additional_information:
          nullable: true
          title: additional_information
          description: additional_information
          type: string
          format: '-'
          maxLength: 4096
        autoapprove:
          nullable: true
          title: autoapprove
          description: autoapprove
          type: string
          format: '-'
          maxLength: 256
      type: object
    SysPermission:
      title: SysPermission
      description: SysPermission
      properties:
        id:
          nullable: false
          title: id
          description: id
          type: integer
          format: '-'
          maxLength: 20
        name:
          nullable: true
          title: name
          description: name
          type: string
          format: '-'
          maxLength: 64
        menu_id:
          nullable: true
          title: menu_id
          description: menu_id
          type: integer
          format: '-'
          maxLength: 11
        url_perm:
          nullable: true
          title: url_perm
          description: url_perm
          type: string
          format: '-'
          maxLength: 128
        btn_perm:
          nullable: true
          title: btn_perm
          description: btn_perm
          type: string
          format: '-'
          maxLength: 64
        gmt_create:
          nullable: true
          title: gmt_create
          description: gmt_create
          type: string
          format: date
        gmt_modified:
          nullable: true
          title: gmt_modified
          description: gmt_modified
          type: string
          format: date
      type: object
...
  requestBodies:
...
    SysOauthClient:
      description: 'SysOauthClient object that needs to be added'
      content:
        application/x-www-form-urlencoded:
          schema:
            $ref: '#/components/schemas/SysOauthClient'
        application/xml:
          schema:
            $ref: '#/components/schemas/SysOauthClient'
        application/json:
          schema:
            $ref: '#/components/schemas/SysOauthClient'
    SysPermission:
      description: 'SysPermission object that needs to be added'
      content:
        application/x-www-form-urlencoded:
          schema:
            $ref: '#/components/schemas/SysPermission'
        application/xml:
          schema:
            $ref: '#/components/schemas/SysPermission'
        application/json:
          schema:
            $ref: '#/components/schemas/SysPermission'
...